api/types/network: move connect/disconnect options types to client module
