+++ api/types/network/network.go
@@ -45,20 +45,6 @@ type CreateOptions struct {
 	Labels     map[string]string // Labels holds metadata specific to the network being created.
 }
 
-// ConnectOptions represents the data to be used to connect a container to the
-// network.
-type ConnectOptions struct {
-	Container      string
-	EndpointConfig *EndpointSettings `json:",omitempty"`
-}
-
-// DisconnectOptions represents the data to be used to disconnect a container
-// from the network.
-type DisconnectOptions struct {
-	Container string
-	Force     bool
-}
-
 // Inspect is the body of the "get network" http response message.
 type Inspect struct {
 	Name       string                      // Name is the name of the network
