+++ client/network_disconnect_opts.go
@@ -1,0 +1,8 @@
+package client
+
+// NetworkDisconnectOptions represents the data to be used to disconnect a container
+// from the network.
+type NetworkDisconnectOptions struct {
+	Container string
+	Force     bool
+}
